Move classes into sub-packages in the menu package
