+++ src/main/java/com/cosimo/utilities/menu/type/PropertyMenu.java
@@ -1,6 +1,5 @@
 package com.cosimo.utilities.menu.type;
 
-import com.cosimo.utilities.menu.AbstractMenu;
 import com.cosimo.utilities.menu.Button;
 import lombok.NonNull;
 import org.bukkit.inventory.Inventory;
